Noise session

+++ LabCalculatorsBusinessLevel/data_access_layer/Models/NoiseCalculatorData.cs
@@ -4,13 +4,13 @@ partial class NoiseCalculatorData
 {
     public int DataId { get; set; }
 
-    public List<float>? SourceLevel { get; set; }
+    public List<double>? SourceLevel { get; set; }
 
-    public List<float>? BackgroundLevel { get; set; }
+    public List<double>? BackgroundLevel { get; set; }
 
-    public List<float>? Delta { get; set; }
+    public List<double>? Delta { get; set; }
 
-    public List<float>? Result { get; set; }
+    public List<double>? Result { get; set; }
 
     public int RegTimeId { get; set; }
